restructure folders for Pages; got jest up and running
