+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "ticket-checkout",
   "private": true,
-  "type": "module",
   "version": "1.0.0",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "test": "jest --watchAll --coverage=false --verbose",
     "preview": "vite preview"
   },
   "browserslist": {
@@ -36,9 +36,12 @@
     "react-error-boundary": "^3.1.4",
     "react-intl": "^6.4.4",
     "react-router-dom": "^6.15.0",
+    "use-http": "^1.0.28",
     "yup": "^1.2.0"
   },
   "devDependencies": {
+    "@swc/core": "^1.3.82",
+    "@swc/jest": "^0.2.29",
     "@testing-library/dom": "^9.3.1",
     "@testing-library/jest-dom": "^6.1.2",
     "@testing-library/react": "^14.0.0",
@@ -64,6 +67,7 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "jest": "^29.6.4",
     "jest-environment-jsdom": "^29.6.4",
+    "jest-fetch-mock": "^3.0.3",
     "jest-transform-stub": "^2.0.0",
     "prettier": "^3.0.3",
     "prettier-plugin-organize-imports": "^3.2.3",
