+++ src/Library/Images/index.ts
@@ -1,9 +1,7 @@
 import Brody from './Brody.png';
 import Carlisle from './Carlisle.png';
-import CreditCards from './CreditCards.png';
 
 export {
   Brody,
   Carlisle, // 159:280
-  CreditCards,
 };
